Move settings observer helpers into focused files
