+++ Packages/CmuxSettings/Sources/CmuxSettings/Stores/UserDefaultsSettingsStore.swift
@@ -134,51 +134,3 @@ public actor UserDefaultsSettingsStore {
         }
     }
 }
-
-private actor SettingObservationState<Value: Sendable & Equatable> {
-    private var lastYielded: Value?
-    private var isRefreshing = false
-    private var needsRefresh = false
-
-    func refresh(
-        read: @escaping @Sendable () async -> Value,
-        to continuation: AsyncStream<Value>.Continuation
-    ) async {
-        if isRefreshing {
-            needsRefresh = true
-            return
-        }
-        isRefreshing = true
-        while true {
-            needsRefresh = false
-            let value = await read()
-            if needsRefresh { continue }
-            if lastYielded != value {
-                lastYielded = value
-                continuation.yield(value)
-            }
-            if needsRefresh { continue }
-            isRefreshing = false
-            return
-        }
-    }
-}
-
-/// Wraps the opaque observer returned by `NotificationCenter.addObserver` so it
-/// can cross task boundaries despite Objective-C not modeling Sendable.
-///
-/// Safety: the token is an immutable reference, this final class exposes no
-/// mutable shared state, and `remove()` only passes that token back to
-/// NotificationCenter's thread-safe observer-removal API. Callers own the
-/// lifecycle and must call `remove()` when their stream terminates.
-final class NotificationObserverToken: @unchecked Sendable {
-    private let token: NSObjectProtocol
-
-    init(_ token: NSObjectProtocol) {
-        self.token = token
-    }
-
-    func remove() {
-        NotificationCenter.default.removeObserver(token)
-    }
-}
